Rename file and require line
