+++ lib/instana/collector.rb
